patch for getdistreader (#172)
* patch getdistreader: convert from `-logL` to `logL` instead of from `chi2` to `logL`

* create new `test_reader.py` file and move over relevant tests from `test_samples.py`; with more readers coming into place I think it is worth having a dedicated test file for this

* make flake8 conform

* only require `almost` equality between GetDist and Anesthetic

* Updated generate getdist data to have -logL instead of chi2 as the second column

* Added tests to improve coverage of reader

* Removed the redundant python2 NameError catch

* return np.arrays instead of lists in `ns_sim`, else passing to `NestedSamples` will throw an error in a <= comparison

* rename `logL` to `minuslogL` in gen_data.py to properly reflect GetDists different data files and to match the terminology in getdistreader.py
