+++ anesthetic/convert.py
@@ -17,7 +17,7 @@ def to_getdist(nested_samples):
     import getdist
     samples = nested_samples.to_numpy()
     weights = nested_samples.weights
-    loglikes = -2*nested_samples.logL.to_numpy()
+    loglikes = -nested_samples.logL.to_numpy()
     names = nested_samples.columns
     ranges = {name: nested_samples._limits(name) for name in names}
     return getdist.mcsamples.MCSamples(samples=samples,
